Exception handling - delete

+++ src/main/java/com/lucas_ldev/spring/services/UserService.java
@@ -2,8 +2,11 @@ package com.lucas_ldev.spring.services;
 
 import com.lucas_ldev.spring.entities.User;
 import com.lucas_ldev.spring.repositories.UserRepository;
+import com.lucas_ldev.spring.services.exceptions.DataBaseException;
 import com.lucas_ldev.spring.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 import java.util.List;
 import java.util.Optional;
@@ -28,7 +31,16 @@ public class UserService {
     }
 
     public void delete(Long id) {
-        repository.deleteById(id);
+        try {
+            if (repository.existsById(id)) {
+                repository.deleteById(id);
+            }
+            else {
+                throw new ResourceNotFoundException(id);
+            }
+        } catch (DataIntegrityViolationException e) {
+            throw new DataBaseException(e.getMessage());
+        }
     }
 
     public User update(Long id, User obj) {
